+++ src/lte/examples/lena-uplink-power-control.cc
@@ -50,7 +50,7 @@ int main (int argc, char *argv[])
   
   Ptr<LteHelper> lteHelper = CreateObject<LteHelper> ();
 
-  uint8_t bandwidth = 25;
+  uint16_t bandwidth = 25;
   double d1 = 0;
 
   // Create Nodes: eNodeB and UE
