+++ src/lte/model/lte-enb-cphy-sap.h
@@ -57,7 +57,7 @@ public:
    * \param ulBandwidth the UL bandwidth in PRBs
    * \param dlBandwidth the DL bandwidth in PRBs
    */
-  virtual void SetBandwidth (uint8_t ulBandwidth, uint8_t dlBandwidth) = 0;
+  virtual void SetBandwidth (uint16_t ulBandwidth, uint16_t dlBandwidth) = 0;
 
   /**
    * \param ulEarfcn the UL EARFCN
@@ -155,7 +155,7 @@ public:
 
   // inherited from LteEnbCphySapProvider
   virtual void SetCellId (uint16_t cellId);
-  virtual void SetBandwidth (uint8_t ulBandwidth, uint8_t dlBandwidth);
+  virtual void SetBandwidth (uint16_t ulBandwidth, uint16_t dlBandwidth);
   virtual void SetEarfcn (uint32_t ulEarfcn, uint32_t dlEarfcn);
   virtual void AddUe (uint16_t rnti);
   virtual void RemoveUe (uint16_t rnti);
@@ -192,7 +192,7 @@ MemberLteEnbCphySapProvider<C>::SetCellId (uint16_t cellId)
 
 template <class C>
 void 
-MemberLteEnbCphySapProvider<C>::SetBandwidth (uint8_t ulBandwidth, uint8_t dlBandwidth)
+MemberLteEnbCphySapProvider<C>::SetBandwidth (uint16_t ulBandwidth, uint16_t dlBandwidth)
 {
   m_owner->DoSetBandwidth (ulBandwidth, dlBandwidth);
 }
