+++ src/lte/model/lte-enb-mac.h
@@ -217,7 +217,7 @@ private:
   * \param ulBandwidth the UL bandwidth
   * \param dlBandwidth the DL bandwidth
   */
-  void DoConfigureMac (uint8_t ulBandwidth, uint8_t dlBandwidth);
+  void DoConfigureMac (uint16_t ulBandwidth, uint16_t dlBandwidth);
   /**
   * \brief Add UE function
   * \param rnti the RNTI
