+++ src/lte/model/lte-spectrum-value-helper.cc
@@ -173,9 +173,9 @@ LteSpectrumValueHelper::GetUplinkCarrierFrequency (uint32_t nUl)
 }
 
 double 
-LteSpectrumValueHelper::GetChannelBandwidth (uint8_t transmissionBandwidth)
+LteSpectrumValueHelper::GetChannelBandwidth (uint16_t transmissionBandwidth)
 {
-  NS_LOG_FUNCTION ((uint16_t) transmissionBandwidth);
+  NS_LOG_FUNCTION (transmissionBandwidth);
   switch (transmissionBandwidth)
     { 
     case 6:
@@ -191,7 +191,7 @@ LteSpectrumValueHelper::GetChannelBandwidth (uint8_t transmissionBandwidth)
     case 100:
       return 20.0e6;
     default:
-      NS_FATAL_ERROR ("invalid bandwidth value " << (uint16_t) transmissionBandwidth);
+      NS_FATAL_ERROR ("invalid bandwidth value " << transmissionBandwidth);
     }
 }
 
@@ -209,7 +209,7 @@ struct LteSpectrumModelId
    */
   LteSpectrumModelId (uint32_t f, uint8_t b);
   uint32_t earfcn; ///< EARFCN
-  uint8_t  bandwidth; ///< bandwidth
+  uint16_t  bandwidth; ///< bandwidth
 };
 
 LteSpectrumModelId::LteSpectrumModelId (uint32_t f, uint8_t b)
@@ -236,9 +236,9 @@ static std::map<LteSpectrumModelId, Ptr<SpectrumModel> > g_lteSpectrumModelMap; 
 
 
 Ptr<SpectrumModel>
-LteSpectrumValueHelper::GetSpectrumModel (uint32_t earfcn, uint8_t txBandwidthConfiguration)
+LteSpectrumValueHelper::GetSpectrumModel (uint32_t earfcn, uint16_t txBandwidthConfiguration)
 {
-  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration);
+  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration);
   Ptr<SpectrumModel> ret;
   LteSpectrumModelId key (earfcn, txBandwidthConfiguration);
   std::map<LteSpectrumModelId, Ptr<SpectrumModel> >::iterator it = g_lteSpectrumModelMap.find (key);
@@ -271,9 +271,9 @@ LteSpectrumValueHelper::GetSpectrumModel (uint32_t earfcn, uint8_t txBandwidthCo
 }
 
 Ptr<SpectrumValue> 
-LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint8_t txBandwidthConfiguration, double powerTx, std::vector <int> activeRbs)
+LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint16_t txBandwidthConfiguration, double powerTx, std::vector <int> activeRbs)
 {
-  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration << powerTx << activeRbs);
+  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration << powerTx << activeRbs);
 
   Ptr<SpectrumModel> model = GetSpectrumModel (earfcn, txBandwidthConfiguration);
   Ptr<SpectrumValue> txPsd = Create <SpectrumValue> (model);
@@ -295,9 +295,9 @@ LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint8_t t
 }
 
 Ptr<SpectrumValue>
-LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint8_t txBandwidthConfiguration, double powerTx, std::map<int, double> powerTxMap, std::vector <int> activeRbs)
+LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint16_t txBandwidthConfiguration, double powerTx, std::map<int, double> powerTxMap, std::vector <int> activeRbs)
 {
-  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration << activeRbs);
+  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration << activeRbs);
 
   Ptr<SpectrumModel> model = GetSpectrumModel (earfcn, txBandwidthConfiguration);
   Ptr<SpectrumValue> txPsd = Create <SpectrumValue> (model);
@@ -333,9 +333,9 @@ LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint8_t t
 }
 
 Ptr<SpectrumValue>
-LteSpectrumValueHelper::CreateUlTxPowerSpectralDensity (uint16_t earfcn, uint8_t txBandwidthConfiguration, double powerTx, std::vector <int> activeRbs)
+LteSpectrumValueHelper::CreateUlTxPowerSpectralDensity (uint16_t earfcn, uint16_t txBandwidthConfiguration, double powerTx, std::vector <int> activeRbs)
 {
-  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration << powerTx << activeRbs);
+  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration << powerTx << activeRbs);
 
   Ptr<SpectrumModel> model = GetSpectrumModel (earfcn, txBandwidthConfiguration);
   Ptr<SpectrumValue> txPsd = Create <SpectrumValue> (model);
@@ -357,9 +357,9 @@ LteSpectrumValueHelper::CreateUlTxPowerSpectralDensity (uint16_t earfcn, uint8_t
 }
 
 Ptr<SpectrumValue>
-LteSpectrumValueHelper::CreateNoisePowerSpectralDensity (uint32_t earfcn, uint8_t txBandwidthConfiguration, double noiseFigure)
+LteSpectrumValueHelper::CreateNoisePowerSpectralDensity (uint32_t earfcn, uint16_t txBandwidthConfiguration, double noiseFigure)
 {
-  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration << noiseFigure);
+  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration << noiseFigure);
   Ptr<SpectrumModel> model = GetSpectrumModel (earfcn, txBandwidthConfiguration);
   return CreateNoisePowerSpectralDensity (noiseFigure, model);
 }
